refactor: extract ExportDestination.swift from IssueExportTargets.swift
Extracts the `ExportDestination` enum (GitHub/Jira destination
identifier) out of IssueExportTargets.swift into its own file so that
file holds target payloads only.

Byte-identical move of lines 89-99. Only imports Foundation.

IssueExportTargets.swift: 124 → 112 lines. Tests: 667.
